requirements.txt
boto3==1.10.40
botocore==1.13.40
docutils==0.15.2
future==0.18.2
javabridge==1.0.18
jmespath==0.9.4
llvmlite==0.30.0
numba==0.46.0
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.1
PyQt5==5.14.0
PyQt5-sip==12.7.0
python-bioformats==1.5.2
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.2.1
SimpleITK==1.2.4
six==1.13.0
urllib3==1.25.7
XlsxWriter==1.2.6
